+++ src/main/java/edu/kis/vh/nursery/FIFORhymer.java
@@ -1,0 +1,25 @@
+package edu.kis.vh.nursery;
+
+public class FIFORhymer extends DefaultCountingOutRhymer {
+
+	private final DefaultCountingOutRhymer temp = new DefaultCountingOutRhymer();
+
+	@Override
+	public int countOut() {
+		while (!callCheck())
+
+		{
+			temp.countIn(super.countOut());
+		}
+
+		int ret = temp.countOut();
+
+		while (!temp.callCheck())
+
+		{
+			countIn(temp.countOut());
+		}
+
+		return ret;
+	}
+}
