--- src/main/java/edu/kis/vh/nursery/FIFORyhmer.java
@@ -1,25 +1,0 @@
-package edu.kis.vh.nursery;
-
-public class FIFORyhmer extends DefaultCountingOutRyhmer {
-
-	private final DefaultCountingOutRyhmer temp = new DefaultCountingOutRyhmer();
-
-	@Override
-	public int countOut() {
-		while (!callCheck())
-
-		{
-			temp.countIn(super.countOut());
-		}
-
-		int ret = temp.countOut();
-
-		while (!temp.callCheck())
-
-		{
-			countIn(temp.countOut());
-		}
-
-		return ret;
-	}
-}
