+++ src/main/java/edu/kis/vh/nursery/HanoiRhymer.java
@@ -1,0 +1,20 @@
+package edu.kis.vh.nursery;
+
+public class HanoiRhymer extends DefaultCountingOutRhymer {
+
+	private int totalRejected = 0;
+
+	public int reportRejected() {
+		return totalRejected;
+	}
+
+	@Override
+	public void countIn(int in) {
+		if (!callCheck() && in > peekaboo()) {
+			totalRejected++;
+		} else {
+			super.countIn(in);
+		}
+	}
+}
+//Kombinacja klawiszy alt+<- lub alt+-> : zmienia położenie kursora na jego poprzednie położenie
