--- src/main/java/edu/kis/vh/nursery/HanoiRyhmer.java
@@ -1,20 +1,0 @@
-package edu.kis.vh.nursery;
-
-public class HanoiRyhmer extends DefaultCountingOutRyhmer {
-
-	private int totalRejected = 0;
-
-	public int reportRejected() {
-		return totalRejected;
-	}
-
-	@Override
-	public void countIn(int in) {
-		if (!callCheck() && in > peekaboo()) {
-			totalRejected++;
-		} else {
-			super.countIn(in);
-		}
-	}
-}
-//Kombinacja klawiszy alt+<- lub alt+-> : zmienia położenie kursora na jego poprzednie położenie
