+++ src/main/java/edu/kis/vh/nursery/factory/DefaultRhymersFactory.java
@@ -1,0 +1,29 @@
+package edu.kis.vh.nursery.factory;
+
+import edu.kis.vh.nursery.DefaultCountingOutRhymer;
+import edu.kis.vh.nursery.FIFORhymer;
+import edu.kis.vh.nursery.HanoiRhymer;
+
+public class DefaultRhymersFactory implements Rhymersfactory {
+
+	@Override
+	public DefaultCountingOutRhymer getStandardRyhmer() {
+		return new DefaultCountingOutRhymer();
+	}
+
+	@Override
+	public DefaultCountingOutRhymer getFalseRyhmer() {
+		return new DefaultCountingOutRhymer();
+	}
+
+	@Override
+	public DefaultCountingOutRhymer getFIFORyhmer() {
+		return new FIFORhymer();
+	}
+
+	@Override
+	public DefaultCountingOutRhymer getHanoiRyhmer() {
+		return new HanoiRhymer();
+	}
+
+}
