+++ src/main/java/edu/kis/vh/nursery/factory/Rhymersfactory.java
@@ -1,0 +1,15 @@
+package edu.kis.vh.nursery.factory;
+
+import edu.kis.vh.nursery.DefaultCountingOutRhymer;
+
+public interface Rhymersfactory {
+
+    DefaultCountingOutRhymer getStandardRyhmer();
+
+    DefaultCountingOutRhymer getFalseRyhmer();
+
+    DefaultCountingOutRhymer getFIFORyhmer();
+
+    DefaultCountingOutRhymer getHanoiRyhmer();
+
+}
