--- src/main/java/edu/kis/vh/nursery/factory/Ryhmersfactory.java
@@ -1,15 +1,0 @@
-package edu.kis.vh.nursery.factory;
-
-import edu.kis.vh.nursery.DefaultCountingOutRyhmer;
-
-public interface Ryhmersfactory {
-
-    DefaultCountingOutRyhmer getStandardRyhmer();
-
-    DefaultCountingOutRyhmer getFalseRyhmer();
-
-    DefaultCountingOutRyhmer getFIFORyhmer();
-
-    DefaultCountingOutRyhmer getHanoiRyhmer();
-
-}
