+++ src/test/java/edu/kis/vh/nursery/RhymersDemo.java
@@ -1,0 +1,37 @@
+package edu.kis.vh.nursery;
+
+import edu.kis.vh.nursery.factory.DefaultRhymersFactory;
+import edu.kis.vh.nursery.factory.Rhymersfactory;
+
+class RhymersDemo {
+
+	public static void main(String[] args) {
+		Rhymersfactory factory = new DefaultRhymersFactory();
+
+		testRhymers(factory);
+		
+	}
+
+	private static void testRhymers(Rhymersfactory factory) {
+		DefaultCountingOutRhymer[] ryhmers = { factory.getStandardRyhmer(), factory.getFalseRyhmer(),
+				factory.getFIFORyhmer(), factory.getHanoiRyhmer()};
+
+		for (int i = 1; i < 15; i++)
+			for (int j = 0; j < 3; j++)
+				ryhmers[j].countIn(i);
+
+		java.util.Random rn = new java.util.Random();
+		for (int i = 1; i < 15; i++)
+			ryhmers[3].countIn(rn.nextInt(20));
+
+		for (int i = 0; i < ryhmers.length; i++) {
+			while (!ryhmers[i].callCheck())
+				System.out.print(ryhmers[i].countOut() + "  ");
+			System.out.println();
+		}
+
+		System.out.println("total rejected is "
+				+ ((HanoiRhymer) ryhmers[3]).reportRejected());
+	}
+
+}
